requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
vispy>=0.14.0
pandas>=1.5.0
PyYAML>=6.0
Pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
requests>=2.31.0
pytube>=15.0.0
pytest>=7.4.0
black>=23.0.0
pylint>=3.0.0
numpy>=1.24.0
trimesh>=3.23.0
